commons package renaming integration
this commit integrates package renaming of the confapi commons
lib which occured with commons version 0.0.8.
