+++ pom.xml
@@ -62,7 +62,7 @@
         <atlassian.gadgets.version>4.3.2.1</atlassian.gadgets.version>
         <atlassian.plugin.key>${project.groupId}.${project.artifactId}</atlassian.plugin.key>
         <atlassian.spring.scanner.version>2.1.5</atlassian.spring.scanner.version>
-        <confapi-commons.version>0.0.7-SNAPSHOT</confapi-commons.version>
+        <confapi-commons.version>0.0.9-SNAPSHOT</confapi-commons.version>
         <hibernate-validator.version>5.2.1.Final</hibernate-validator.version>
         <javax.el-api.version>2.2.4</javax.el-api.version>
         <junit.version>4.12</junit.version>
@@ -101,7 +101,7 @@
         <!-- compile dependencies -->
 
         <dependency>
-            <groupId>de.aservo.atlassian</groupId>
+            <groupId>de.aservo</groupId>
             <artifactId>confapi-commons</artifactId>
             <version>${confapi-commons.version}</version>
             <scope>compile</scope>
@@ -213,7 +213,7 @@
         <!-- test dependencies -->
 
         <dependency>
-            <groupId>de.aservo.atlassian</groupId>
+            <groupId>de.aservo</groupId>
             <artifactId>confapi-commons</artifactId>
             <version>${confapi-commons.version}</version>
             <type>test-jar</type>
