+++ src/main/java/de/aservo/atlassian/confluence/confapi/model/util/UserBeanUtil.java
@@ -2,7 +2,7 @@ package de.aservo.atlassian.confluence.confapi.model.util;
 
 import com.atlassian.confluence.user.ConfluenceUserImpl;
 import com.atlassian.user.User;
-import de.aservo.atlassian.confapi.model.UserBean;
+import de.aservo.confapi.commons.model.UserBean;
 
 public class UserBeanUtil {
