+++ src/main/java/de/aservo/atlassian/confluence/confapi/rest/PermissionsResourceImpl.java
@@ -5,11 +5,11 @@ import com.atlassian.confluence.security.SpacePermissionManager;
 import com.atlassian.confluence.security.service.AnonymousUserPermissionsService;
 import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
 import com.sun.jersey.spi.container.ResourceFilters;
-import de.aservo.atlassian.confapi.constants.ConfAPI;
-import de.aservo.atlassian.confapi.model.ErrorCollection;
 import de.aservo.atlassian.confluence.confapi.filter.AdminOnlyResourceFilter;
 import de.aservo.atlassian.confluence.confapi.model.PermissionAnonymousAccessBean;
 import de.aservo.atlassian.confluence.confapi.rest.api.PermissionsResource;
+import de.aservo.confapi.commons.constants.ConfAPI;
+import de.aservo.confapi.commons.model.ErrorCollection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
