+++ src/main/java/de/aservo/atlassian/confluence/confapi/service/GadgetsServiceImpl.java
@@ -9,10 +9,10 @@ import com.atlassian.gadgets.directory.spi.ExternalGadgetSpecStore;
 import com.atlassian.gadgets.spec.GadgetSpecFactory;
 import com.atlassian.plugin.spring.scanner.annotation.export.ExportAsService;
 import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
-import de.aservo.atlassian.confapi.exception.BadRequestException;
-import de.aservo.atlassian.confapi.model.GadgetBean;
-import de.aservo.atlassian.confapi.model.GadgetsBean;
-import de.aservo.atlassian.confapi.service.api.GadgetsService;
+import de.aservo.confapi.commons.exception.BadRequestException;
+import de.aservo.confapi.commons.model.GadgetBean;
+import de.aservo.confapi.commons.model.GadgetsBean;
+import de.aservo.confapi.commons.service.api.GadgetsService;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
