+++ src/test/java/de/aservo/atlassian/confluence/confapi/model/util/DirectoryBeanUtilTest.java
@@ -3,7 +3,7 @@ package de.aservo.atlassian.confluence.confapi.model.util;
 import com.atlassian.crowd.embedded.api.Directory;
 import com.atlassian.crowd.embedded.api.DirectoryType;
 import com.atlassian.crowd.model.directory.DirectoryImpl;
-import de.aservo.atlassian.confapi.model.DirectoryBean;
+import de.aservo.confapi.commons.model.DirectoryBean;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.runners.MockitoJUnitRunner;
