+++ src/test/java/de/aservo/atlassian/confluence/confapi/model/util/MailServerSmtpBeanUtilTest.java
@@ -2,7 +2,7 @@ package de.aservo.atlassian.confluence.confapi.model.util;
 
 import com.atlassian.mail.server.DefaultTestSmtpMailServerImpl;
 import com.atlassian.mail.server.SMTPMailServer;
-import de.aservo.atlassian.confapi.model.MailServerSmtpBean;
+import de.aservo.confapi.commons.model.MailServerSmtpBean;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.runners.MockitoJUnitRunner;
