+++ src/test/java/de/aservo/atlassian/confluence/confapi/rest/PermissionsResourceTest.java
@@ -3,8 +3,8 @@ package de.aservo.atlassian.confluence.confapi.rest;
 import com.atlassian.confluence.security.SpacePermission;
 import com.atlassian.confluence.security.SpacePermissionManager;
 import com.atlassian.confluence.security.service.AnonymousUserPermissionsService;
-import de.aservo.atlassian.confapi.model.ErrorCollection;
 import de.aservo.atlassian.confluence.confapi.model.PermissionAnonymousAccessBean;
+import de.aservo.confapi.commons.model.ErrorCollection;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
