+++ src/test/java/de/aservo/atlassian/confluence/confapi/service/ApplicationLinkServiceTest.java
@@ -9,11 +9,11 @@ import com.atlassian.applinks.spi.manifest.ManifestNotFoundException;
 import com.atlassian.applinks.spi.util.TypeAccessor;
 import com.atlassian.confluence.settings.setup.DefaultApplicationLink;
 import com.atlassian.confluence.settings.setup.DefaultApplicationType;
-import de.aservo.atlassian.confapi.model.ApplicationLinkBean;
-import de.aservo.atlassian.confapi.model.ApplicationLinksBean;
-import de.aservo.atlassian.confapi.model.type.ApplicationLinkTypes;
 import de.aservo.atlassian.confluence.confapi.model.DefaultAuthenticationScenario;
 import de.aservo.atlassian.confluence.confapi.model.util.ApplicationLinkBeanUtil;
+import de.aservo.confapi.commons.model.ApplicationLinkBean;
+import de.aservo.confapi.commons.model.ApplicationLinksBean;
+import de.aservo.confapi.commons.model.type.ApplicationLinkTypes;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
