+++ src/test/java/de/aservo/atlassian/confluence/confapi/service/DirectoryServiceTest.java
@@ -5,9 +5,9 @@ import com.atlassian.crowd.embedded.api.Directory;
 import com.atlassian.crowd.embedded.api.DirectoryType;
 import com.atlassian.crowd.exception.DirectoryCurrentlySynchronisingException;
 import com.atlassian.crowd.model.directory.ImmutableDirectory;
-import de.aservo.atlassian.confapi.exception.InternalServerErrorException;
-import de.aservo.atlassian.confapi.model.DirectoryBean;
 import de.aservo.atlassian.confluence.confapi.model.util.DirectoryBeanUtil;
+import de.aservo.confapi.commons.exception.InternalServerErrorException;
+import de.aservo.confapi.commons.model.DirectoryBean;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
