+++ src/test/java/de/aservo/atlassian/confluence/confapi/service/LicensesServiceTest.java
@@ -3,10 +3,10 @@ package de.aservo.atlassian.confluence.confapi.service;
 import com.atlassian.confluence.settings.setup.DefaultSingleProductLicenseDetailsView;
 import com.atlassian.sal.api.i18n.InvalidOperationException;
 import com.atlassian.sal.api.license.LicenseHandler;
-import de.aservo.atlassian.confapi.exception.BadRequestException;
-import de.aservo.atlassian.confapi.exception.InternalServerErrorException;
-import de.aservo.atlassian.confapi.model.LicenseBean;
-import de.aservo.atlassian.confapi.model.LicensesBean;
+import de.aservo.confapi.commons.exception.BadRequestException;
+import de.aservo.confapi.commons.exception.InternalServerErrorException;
+import de.aservo.confapi.commons.model.LicenseBean;
+import de.aservo.confapi.commons.model.LicensesBean;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
