+++ src/test/java/de/aservo/atlassian/confluence/confapi/service/MailServerServiceTest.java
@@ -8,11 +8,11 @@ import com.atlassian.mail.server.OtherTestPopMailServerImpl;
 import com.atlassian.mail.server.OtherTestSmtpMailServerImpl;
 import com.atlassian.mail.server.PopMailServer;
 import com.atlassian.mail.server.SMTPMailServer;
-import de.aservo.atlassian.confapi.exception.BadRequestException;
-import de.aservo.atlassian.confapi.model.MailServerPopBean;
-import de.aservo.atlassian.confapi.model.MailServerSmtpBean;
 import de.aservo.atlassian.confluence.confapi.model.util.MailServerPopBeanUtil;
 import de.aservo.atlassian.confluence.confapi.model.util.MailServerSmtpBeanUtil;
+import de.aservo.confapi.commons.exception.BadRequestException;
+import de.aservo.confapi.commons.model.MailServerPopBean;
+import de.aservo.confapi.commons.model.MailServerSmtpBean;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
