+++ src/test/java/de/aservo/atlassian/confluence/confapi/service/UserServiceTest.java
@@ -4,9 +4,9 @@ import com.atlassian.user.EntityException;
 import com.atlassian.user.User;
 import com.atlassian.user.UserManager;
 import com.atlassian.user.impl.DefaultUser;
-import de.aservo.atlassian.confapi.exception.BadRequestException;
-import de.aservo.atlassian.confapi.exception.NotFoundException;
-import de.aservo.atlassian.confapi.model.UserBean;
+import de.aservo.confapi.commons.exception.BadRequestException;
+import de.aservo.confapi.commons.exception.NotFoundException;
+import de.aservo.confapi.commons.model.UserBean;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
