+++ src/test/java/it/de/aservo/atlassian/confluence/confapi/rest/LicenseResourceTest.java
@@ -1,5 +1,5 @@
 package it.de.aservo.atlassian.confluence.confapi.rest;
 
-import it.de.aservo.atlassian.confapi.rest.AbstractLicenseResourceFuncTest;
+import it.de.aservo.confapi.commons.rest.AbstractLicenseResourceFuncTest;
 
 public class LicenseResourceTest extends AbstractLicenseResourceFuncTest { }
