+++ src/test/java/it/de/aservo/atlassian/confluence/confapi/rest/PingResourceTest.java
@@ -1,5 +1,5 @@
 package it.de.aservo.atlassian.confluence.confapi.rest;
 
-import it.de.aservo.atlassian.confapi.rest.AbstractPingResourceFuncTest;
+import it.de.aservo.confapi.commons.rest.AbstractPingResourceFuncTest;
 
 public class PingResourceTest extends AbstractPingResourceFuncTest { }
